# perf(crypto): use ring for asm implementations of sha256/sha512